src/third_party/WebKit/LayoutTests/fast/html/imports/import-same-url-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
PASS target1.import is non-null.
PASS target1.import is target2.import
PASS target1.import is target3.import
PASS successfullyParsed is true

TEST COMPLETE
